requirements.txt
python-telegram-bot==21.6
faster-whisper==1.0.3
ffmpeg-python==0.2.0
pydub==0.25.1